refactor: extract makeTransferRoute tests
